refactor(record): consolidate rally-related files and standardize naming
- Combine create-rally.usecase.ts and update-rally.usecase.ts into rally.usecase.ts
- Merge create-rally.controller.ts and update-rally.controller.ts into rally.controller.ts
- Rename helper functions from *Optimistic to *Helper for consistency
  - createRallyOptimistic → createRallyHelper
  - updateRallyOptimistic → updateRallyHelper
  - createSubstitutionOptimistic → createSubstitutionHelper
- Update all import paths and tests to reference new structure
- Improve code organization by grouping related functionality
- Remove redundant object assignments in helper functions

This change establishes a resource-based file naming convention instead of
action-based naming, improving consistency across the codebase.
