+++ src/interface/controllers/record/rally.controller.ts
@@ -1,10 +1,23 @@
 import { container } from "@/infrastructure/di/inversify.config";
 import { TYPES } from "@/infrastructure/di/types";
 import {
+  CreateRallyUseCase,
+  type ICreateRallyInput,
+  type ICreateRallyOutput,
   UpdateRallyUseCase,
   type IUpdateRallyInput,
   type IUpdateRallyOutput,
-} from "@/applications/usecases/record/update-rally.usecase";
+} from "@/applications/usecases/record/rally.usecase";
+
+export const createRallyController = async (
+  input: ICreateRallyInput
+): Promise<ICreateRallyOutput | undefined> => {
+  const createRallyUseCase = container.get<CreateRallyUseCase>(
+    TYPES.CreateRallyUseCase
+  );
+
+  return await createRallyUseCase.execute(input);
+};
 
 export const updateRallyController = async (
   input: IUpdateRallyInput
